pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.version]
path = "whatamithinking/jsonproto/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["whatamithinking.jsonproto"]

[project]
authors = [
  {name = "Connor Sherwood Maynes", email = "connormaynes@gmail.com"},
]
classifiers = []
dependencies = [
  "orjson",
  "lru-dict",
]
description = "declarative json modeler, validator, encoder, and decoder"
dynamic = ["version", "readme"]
keywords = []
maintainers = []
name = "whatamithinking-jsonproto"
requires-python = ">=3.12"

[tool.setuptools.dynamic]
readme = {file = "README.md"}
version = {attr = "whatamithinking.jsonproto.__version__"}

[project.optional-dependencies]
dev = ["black"]
test = []
benchmarks = [
  "pydantic>=2",
  "pyserde",
  "attrs",
  "mashumaro",
  "dataclassy",
  "msgspec",
  "psutil",
  "apischema",
]

[project.scripts]
[project.gui-scripts]
